# Adds error for mixing NumPy and CVXPY