Following up on #1647 

This code is supposed to not raise an error on anything that used to succeed (which was just the `np.hstack`/`np.vstack` equivalents I believe); however, by adding the `__array_function__` method NumPy is now assuming we can handle everything and when we say we can't (i.e. we return `NotImplemented`) it chooses to errors instead of creating an array of objects as it used to. The error message has changed on a number of things to say that we have a `TypeError` instead of an `AttributeError`.

If breaking code that constructs arrays with `np.hstack` and `np.vstack` is not acceptable, (`np.array` will still work though) then we can probably just eliminate the `__array_function__` implementation, but I don't see any easy way to generate the warning while preserving old behavior unless we want to do `AutoDiff`-style manual implementation of each thing we want to keep working.