@SteveDiamond 2022-02-09
Actually, if it's only the ufunc left, it seems fine to raise an error.  Make a new test file ``test_numpy.py``.