@PTNobel 2022-02-28
Hey @SteveDiamond , I can def get it done this week. I had some emergency personal stuff pop up and also took my quals last week, so I got a bit distracted. I'm doing the 364A final on Wednesday, and then should be able to focus on this.