@PTNobel 2022-03-05
Okay, so I had forgotten that NumPy prefers to call an `__array_ufunc__` method over `__radd__` for code of the form `np.ndarray + cp.Variable`. Worse, if we return NotImplemented NumPy raises a TypeError rather than pass control back to Python so it can try `__radd__`.

Here's my idea for a solution:

 * Have a known set of ufuncs that correspond to binary operations that we don't raise an error for (i.e. `np.left_shift`, `np.right_shift`, `np.add`, `np.subtract`, `np.divide`, `np.multiply`, `np.matmul`, `np.power`, am I missing any?)
 * If args is not an instance of [np.ndarray, cp.Expression] or kwargs is not {} (i.e. we're not in the case triggered by the binary operators), then we raise the exception as normal.
 * If we are in that case, we just manually make the appropriate call, i.e. `self.__radd__(array, expr)`.

Thoughts?