@PTNobel 2022-03-05
I just pushed an implementation of my proposal.

I have a tiny test script I'm using right now; I'll write some more careful tests in the CVXPY testing framework, once I have the go-ahead on this plan.

(Also I might be triggering a new warning on `ndarray == expression`, I'm not sure why, since it isn't calling into `__array_ufunc__`)

```
import cvxpy as cp
import numpy as np

x = cp.Variable()
try:
    np.sqrt(x)
except RuntimeError:
    pass
else:
    assert False, "np.sqrt(cp.Variable()) Should have raised an error"

try:
    x + np.array([1])
except RuntimeError:
    assert False, "cp.Variable() + np.ndarray() should not have raised an error"
else:
    pass

try:
    np.array([1]) + x
except RuntimeError:
    print("np.ndarray() + cp.Variable() raised an error")
else:
    pass

try:
    x == np.array([1])
except RuntimeError:
    assert False, "cp.Variable() == np.ndarray() should not have raised an error"
else:
    pass

try:
    np.array([1]) == x
except RuntimeError:
    assert False, "np.ndarray() == cp.Variable() should not have raised an error"
else:
    pass

```