@PTNobel 2022-03-05
Ah I figured out the issue. NumPy was calling our `__array_ufunc__` with np.equal, but catching the error, raising a deprecation warning and then falling back to the old behaviour. All we have to do is add `np.equal` to the list and we'll get the behaviour we want.