@PTNobel 2022-03-05
Alright, i fixed that bug; since I now handle every binary operator (the only places NumPy might start trying to call this function without the user naming a NumPy function), I think this should have feature parity with the old code!

I'm going through the Contributing checklist and figuring out how to run the tests so I have some more confidence on that though.