@PTNobel 2022-03-05
Well, it turns out numpy types other than ndarray (e.g. int64, float64) also call __array_ufunc__ to implement multiplication. My gut feeling is to just remove the "check if the first argument is a np.ndarray" step in my algorithm, so that any type on the left of a binary operator works correctly.