@PTNobel 2022-03-05
Okay, I fixed a couple minor bugs and now have:

    751 passed, 182 skipped, 330 warnings in 64.86s (0:01:04) 

I'm a bit nervous there will be a performance impact to have a dictionary lookup on every add and subtract, so I'm going to run those tests now.