@PTNobel 2022-03-05
With my patch:

```
===================== test session starts =====================
platform linux -- Python 3.10.2, pytest-7.0.1, pluggy-1.0.0
rootdir: /home/parthnobel/Research/cvxpy, configfile: pyproject.toml
collected 11 items

test_benchmarks.py cone_matrix_stuffing_with_many_constraints: avg=1.339e+00 s , std=0.000e+00 s (1 iterations)
.diffcp_sdp: avg=5.383e+00 s , std=0.000e+00 s (1 iterations)
.least_squares: avg=6.506e-03 s , std=0.000e+00 s (1 iterations)
.sConic canonicalization
(ECOS) solver time:  0.061479839
cvxpy time:  0.6716441787154541
Quadratic canonicalization
(OSQP) solver time:  0.031092942
cvxpy time:  0.7756702301496582
.qp: avg=1.731e-02 s , std=0.000e+00 s (1 iterations)
.small_cone_matrix_stuffing: avg=4.953e-02 s , std=1.317e-02 s (10 iterations)
.small_lp: avg=2.414e-02 s , std=0.000e+00 s (1 iterations)
small_lp_second_time: avg=2.008e-03 s , std=0.000e+00 s (1 iterations)
.sstv_inpainting: avg=2.752e+00 s , std=0.000e+00 s (1 iterations)
.
```

Without my patch:

```
===================================================== test session starts =====================================================
platform linux -- Python 3.10.2, pytest-7.0.1, pluggy-1.0.0
rootdir: /home/parthnobel/Research/cvxpy, configfile: pyproject.toml
collected 11 items

test_benchmarks.py cone_matrix_stuffing_with_many_constraints: avg=1.273e+00 s , std=0.000e+00 s (1 iterations)
.diffcp_sdp: avg=5.556e+00 s , std=0.000e+00 s (1 iterations)
.least_squares: avg=7.001e-03 s , std=0.000e+00 s (1 iterations)
.sConic canonicalization
(ECOS) solver time:  0.060723004
cvxpy time:  0.6589372781350098
Quadratic canonicalization
(OSQP) solver time:  0.030688248
cvxpy time:  0.7772355457927246
.qp: avg=1.679e-02 s , std=0.000e+00 s (1 iterations)
.small_cone_matrix_stuffing: avg=4.559e-02 s , std=1.042e-02 s (10 iterations)
.small_lp: avg=2.286e-02 s , std=0.000e+00 s (1 iterations)
small_lp_second_time: avg=1.629e-03 s , std=0.000e+00 s (1 iterations)
.sstv_inpainting: avg=2.704e+00 s , std=0.000e+00 s (1 iterations)
.

```