@PTNobel 2022-03-05
Fixed a formatting issue! Let me know if you all can tolerate the slowdown and like my approach here, I'll squash these commits and rebase onto master and then add a second commit with tests.