@rileyjmurray 2022-03-05
@PTNobel can you summarize your changes? I have two specific questions: 
1. What happens with calling ``np.hstack`` and ``np.vstack`` on cvxpy Expressions? I think the current behavior is useful because it provides a way for users to implement sophisticated affine transformations which cvxpy might not natively support.
2.The dictionary lookup is only used when one of the operands in a binary operation is a numpy datatype, right? If so, then I think the performance impact should be relatively modest. We encourage users to build vectorized models, so it's not like they should have several thousand operations which hit that code path.