@SteveDiamond 2022-03-05
I'm in the middle of some stuff and can't read all the comments. However, it appears that this changes may be more complicated than we had thought, with more unexpected consequences. I would suggest we hold off on adding this to 1.2.